Refactor file setup
